+++ R/mutual_difference.R
@@ -1,8 +1,9 @@
 #' Decomposes the difference between two M indices
 #'
 #' Uses one of three methods to decompose the difference between two M indices:
-#' (1) "shapley": a method based on the Shapley decomposition with a few advantages over the
-#' Karmel-Maclachlan method (recommended and the default, Deutsch et al. 2006),
+#' (1) "shapley" / "shapley_detailed": a method based on the Shapley decomposition
+#' with a few advantages over the Karmel-Maclachlan method
+#' (recommended and the default, Deutsch et al. 2006),
 #' (2) "km": the method based on Karmel-Maclachlan (1988),
 #' (3) "mrc": the method developed by Mora and Ruiz-Castillo (2009).
 #' All methods have been extended to account for missing units/groups in either data input.
@@ -10,6 +11,8 @@
 #' The Shapley method is an improvement over the Karmel-Maclachlan method (Deutsch et al. 2006).
 #' It is based on several margins-adjusted data inputs
 #' and yields symmetrical results (i.e. \code{data1} and \code{data2} can be switched).
+#' When "shapley_detailed" is used, the structural component is further decomposed into
+#' the contributions of individuals units.
 #'
 #' The Karmel-Maclachlan method (Karmel and Maclachlan 1988) adjusts
 #' the margins of \code{data1} to be similar to the margins of \code{data2}. This process
@@ -41,23 +44,22 @@
 #' @param unit A categorical variable or a vector of variables
 #'   contained in \code{data}. Defines the second dimension
 #'   over which segregation is computed.
-#' @param weight Numeric. Only frequency weights are allowed.
-#'   (Default \code{NULL})
+#' @param weight Numeric. (Default \code{NULL})
 #' @param method Either "shapley" (the default), "km" (Karmel and Maclachlan method), or
 #'   "mrc" (Mora and Ruiz-Castillo method).
 #' @param se If \code{TRUE}, standard errors are estimated via bootstrap.
 #'   (Default \code{FALSE})
-#' @param n_bootstrap Number of bootstrap iterations. (Default \code{50})
+#' @param n_bootstrap Number of bootstrap iterations. (Default \code{100})
 #' @param ... Only used for additional arguments when
 #'  when \code{method} is set to \code{shapley} or \code{km}. See \link{ipf} for details.
 #' @param base Base of the logarithm that is used in the calculation.
 #'   Defaults to the natural logarithm.
-#' @return Returns a data frame with columns \code{stat} and \code{est}. The data frame contains
+#' @return Returns a data.table with columns \code{stat} and \code{est}. The data frame contains
 #'   the following rows defined by \code{stat}:
 #'   \code{M1} contains the M for \code{data1}.
 #'   \code{M2} contains the M for \code{data2}.
 #'   \code{diff} is the difference between \code{M2} and \code{M1}.
-#'   The sum of all rows following \code{diff} equal \code{diff}.
+#'   The sum of the five rows following \code{diff} equal \code{diff}.
 #'
 #'   \code{additions} contains the change in M induces by \code{unit} and \code{group} categories
 #'   present in \code{data2} but not \code{data1}, and \code{removals} the reverse.
@@ -71,6 +73,17 @@
 #'   When using "km", one additional row is returned:
 #'    \code{interaction} is the contribution of differences in the joint marginal distribution
 #'      of \code{unit} and \code{group}.
+#'
+#'   When "shapley_detailed" is used, an additional column "unit" is returned, along with
+#'     five additional rows for each unit that is present in both \code{data1} and \code{data2}.
+#'     The five rows have the following meaning:
+#'     \code{p1} (\code{p2}) is the proportion of the unit in \code{data1} (\code{data2})
+#'     once non-intersecting units/groups have been removed. The changes in local linkage are
+#'     given by \code{ls_diff1} and \code{ls_diff2}. The row named \code{total}
+#'     summarizes the contribution of
+#'     the unit towards structural change
+#'     using the formula \code{.5 * p1 * ls_diff1 + .5 * p2 * ls_diff2}.
+#'     The sum of all "total" components equals structural change.
 #'
 #'   If \code{se} is set to \code{TRUE}, an additional column \code{se} contains
 #'   the associated bootstrapped standard errors, and the column \code{est} contains
@@ -117,13 +130,19 @@
 #' @export
 mutual_difference <- function(data1, data2, group, unit,
                               weight = NULL, method = "shapley",
-                              se = FALSE, n_bootstrap = 50, base = exp(1), ...) {
+                              se = FALSE, n_bootstrap = 100, base = exp(1), ...) {
     if (method == "shapley") {
-        method <- function(...) shapley_compute(...)
+        fun <- function(...) shapley_compute(..., detail = FALSE)
+        cols <- "stat"
+    } else if (method == "shapley_detailed") {
+        fun <- function(...) shapley_compute(..., detail = TRUE)
+        cols <- c("stat", unit)
     } else if (method == "km") {
-        method <- function(...) km_compute(...)
+        fun <- function(...) km_compute(...)
+        cols <- "stat"
     } else if (method == "mrc") {
-        method <- function(...) mrc_compute(...)
+        fun <- function(...) mrc_compute(...)
+        cols <- "stat"
     } else {
         stop("unknown decomposition method")
     }
@@ -139,15 +158,17 @@ mutual_difference <- function(data1, data2, group, unit,
     if (nrow_unit == 0) stop("No overlap in unit")
 
     if (se == FALSE) {
-        ret <- method(d1, d2, group, unit, base, ...)
+        ret <- fun(d1, d2, group, unit, base, ...)
     } else {
         vars <- attr(d1, "vars")
         n_total1 <- sum(d1[, "freq"])
         n_total2 <- sum(d2[, "freq"])
 
-        if (!all(d1[["freq"]] == round(d1[["freq"]])) |
-            !all(d2[["freq"]] == round(d2[["freq"]]))) {
-            warning("bootstrap with non-integer weights")
+        if (all.equal(n_total1, round(n_total1)) != TRUE |
+            all.equal(n_total2, round(n_total2)) != TRUE) {
+            stop(paste0(
+                "bootstrap with a total sample size that is not an integer is not allowed, ",
+                "maybe scale your weights?"))
         }
 
         boot_ret <- lapply(1:n_bootstrap, function(i) {
@@ -160,20 +181,19 @@ mutual_difference <- function(data1, data2, group, unit,
             resampled2 <- d2[
                 sample(.N, n_total2, replace = TRUE, prob = freq)][,
                 list(freq = as.double(.N)), by = vars]
-            method(resampled1, resampled2, group, unit, base, ...)
+            fun(resampled1, resampled2, group, unit, base, ...)
         })
         boot_ret <- rbindlist(boot_ret)
         # summarize bootstrapped data frames
         ret <- boot_ret[, list(
-            est = mean(est), se = stats::sd(est)), by = c("stat")]
+            est = mean(est), se = stats::sd(est)), by = cols]
     }
     close_log()
-    rownames(ret) <- ret[["stat"]]
-    as_df(ret)
+    ret
 }
 
 
-shapley_compute <- function(d1, d2, group, unit, base, ...) {
+shapley_compute <- function(d1, d2, group, unit, base, detail, ...) {
     m <- function(d, weight) {
         add_local(d, group, unit, base, weight)
         sum(d[, list(p = first(p_unit), ls = first(ls_unit)), by = unit][, p * ls])
@@ -188,29 +208,30 @@ shapley_compute <- function(d1, d2, group, unit, base, ...) {
         "n_unit", "n_group", "ls_unit") := NULL]
 
     # group = A, unit = A, structure = A
-    d_AAA <- create_common_data(d1, d2, group, unit)
+    d_AAA <- create_common_data(d1, d2, group, unit, suppress_warnings = TRUE)
     # group = B, unit = B, structure = B
     d_BBB <- copy(d_AAA)
     setnames(d_BBB, c("freq1", "freq2"), c("freq2", "freq1"))
+    setnames(d_BBB, c("freq_orig1", "freq_orig2"), c("freq_orig2", "freq_orig1"))
 
     update_log(ipf_n = 1, ipf_max = 6)
-    d_BBA <- ipf_compute(d_AAA, group, unit, ...)
+    d_BBA <- ipf_compute(copy(d_AAA), group, unit, ...)
     update_log(ipf_n = 2, ipf_max = 6)
-    d_ABA <- ipf_compute(d_AAA, group, unit, only_unit = TRUE, ...)
+    d_ABA <- ipf_compute(copy(d_AAA), group, unit, only_unit = TRUE, ...)
     update_log(ipf_n = 3, ipf_max = 6)
-    d_BAA <- ipf_compute(d_AAA, group, unit, only_group = TRUE, ...)
+    d_BAA <- ipf_compute(copy(d_AAA), group, unit, only_group = TRUE, ...)
 
     update_log(ipf_n = 4, ipf_max = 6)
-    d_AAB <- ipf_compute(d_BBB, group, unit, ...)
+    d_AAB <- ipf_compute(copy(d_BBB), group, unit, ...)
     update_log(ipf_n = 5, ipf_max = 6)
-    d_BAB <- ipf_compute(d_BBB, group, unit, only_unit = TRUE, ...)
+    d_BAB <- ipf_compute(copy(d_BBB), group, unit, only_unit = TRUE, ...)
     update_log(ipf_n = 6, ipf_max = 6)
-    d_ABB <- ipf_compute(d_BBB, group, unit, only_group = TRUE, ...)
+    d_ABB <- ipf_compute(copy(d_BBB), group, unit, only_group = TRUE, ...)
 
     # compute M based on unadjusted counts (where zeros are not replaced by small mumbers)
     # this is to ensure that M2 - m_BBB is exactly 0 when there are no additions
     m_AAA <- m(d_AAA, "freq_orig1")
-    m_BBB <- m(d_AAA, "freq_orig2")
+    m_BBB <- m(d_BBB, "freq_orig1")
     m_BBA <- m(d_BBA, "n")
     m_ABA <- m(d_ABA, "n")
     m_BAA <- m(d_BAA, "n")
@@ -226,13 +247,31 @@ shapley_compute <- function(d1, d2, group, unit, base, ...) {
     unit_marginal <- .25 * (m_ABA - m_AAA + (m_BBA - m_BAA) +
                             m_BBB - m_BAB + (m_ABB - m_AAB))
     stopifnot(round(group_marginal + unit_marginal, 4) == round(marginal, 4))
+    stopifnot(round(M2 - m_BBB + m_AAA - M1 + marginal + structural, 4) == round(M2 - M1, 4))
 
     stat <- c("M1", "M2", "diff", "additions", "removals",
              "group_marginal", "unit_marginal", "structural")
     est <- c(M1, M2, M2 - M1, M2 - m_BBB, m_AAA - M1,
              group_marginal, unit_marginal, structural)
 
-    data.table(stat = stat, est = est, stringsAsFactors = FALSE)
+    ret <- data.table(stat = stat, est = est, stringsAsFactors = FALSE)
+    if (detail == TRUE) {
+        ls_AAB <- d_AAB[, list(ls_AAB = first(ls_unit)), by = unit]
+        ls_AAA <- d_AAA[, list(p1 = first(p_unit), ls_AAA = first(ls_unit)), by = unit]
+        ls_BBB <- d_BBB[, list(p2 = first(p_unit), ls_BBB = first(ls_unit)), by = unit]
+        ls_BBA <- d_BBA[, list(ls_BBA = first(ls_unit)), by = unit]
+        ls <- Reduce(merge, list(ls_AAB, ls_AAA, ls_BBB, ls_BBA))
+        ls[, `:=`(ls_diff1 = ls_AAB - ls_AAA, ls_diff2 = ls_BBB - ls_BBA)]
+        ls[, c("ls_AAB", "ls_AAA", "ls_BBB", "ls_BBA") := NULL]
+        ls[, total := .5 * p1 * ls_diff1 + .5 * p2 * ls_diff2]
+        ls <- melt(ls, unit, variable.name = "stat", value.name = "est")
+        setorderv(ls, unit)
+        ret <- rbindlist(list(ret, ls), fill = TRUE)
+        cols <- c("stat", unit, "est")
+        ret[, cols, with = FALSE]
+    } else {
+        ret
+    }
 }
 
 km_compute <- function(d1, d2, group, unit, base, ...) {
@@ -248,7 +287,7 @@ km_compute <- function(d1, d2, group, unit, base, ...) {
     d2[, c("p_unit", "p_group", "p_group_g_unit",
         "n_unit", "n_group", "ls_unit") := NULL]
 
-    d1_source <- create_common_data(d1, d2, group, unit)
+    d1_source <- create_common_data(d1, d2, group, unit, suppress_warnings = TRUE)
 
     # compute M based on unadjusted counts (where zeros are not replaced by small mumbers)
     # this is to ensure that M2 - m_BBB is exactly 0 when there are no additions
@@ -257,7 +296,7 @@ km_compute <- function(d1, d2, group, unit, base, ...) {
 
     # compute for dataset 1, but with marginals from dataset 2
     update_log(ipf_n = 1, ipf_max = 1)
-    d1_marg_d2 <- ipf_compute(d1_source, group, unit, ...)
+    d1_marg_d2 <- ipf_compute(copy(d1_source), group, unit, ...)
     d1_marg_d2[, `:=`(n_group = sum(n_source), n_group_target = sum(n_target)), by = group]
     d1_marg_d2[, `:=`(n_unit = sum(n_source), n_unit_target = sum(n_target)), by = unit]
     d1_marg_d2[, `:=`(
@@ -293,7 +332,7 @@ mrc_compute <- function(d1, d2, group, unit, base) {
     M2 <- m(d2)
 
     # reduce d1 and d2 to the joint set
-    joined <- create_common_data(d1, d2, group, unit, fill_na = 0)
+    joined <- create_common_data(d1, d2, group, unit, suppress_warnings = TRUE, fill_na = 0)
 
     # reduce M for d1
     add_local(joined, group, unit, base, "freq1")